Move all timezone logic and state to the timezone input component
